+++ functions/api/admin/[city]/debug-feed.ts
@@ -1,5 +1,5 @@
-import { Env } from "../../_core/types";
-import { withCityRoute } from "../../_core/api-utils";
+import { Env } from "../../../_core/types";
+import { withCityRoute } from "../../../_core/api-utils";
 
 export const onRequest: PagesFunction<Env> = withCityRoute(
     (adapter, context) => {